Add naive snapcraft.yaml
This patch moves all of the old content, except for licenses and meta/
into the pre-built directory.  The snapcraft.yaml is derived from the
old meta/snap.yaml with the addition of the required confinement and
grade properties (as suggested by snapcraft). All of the old pre-built
content is installed with a pre-built part using the dump plugin which
simply copies the files across.

The icon is moved to the setup/gui directory, after building it ends up
in a different location but arguably the one that is more correct
(meta/gui/icon.png).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>