snapcraft.yaml: use "on amd64" to install gcc-arm-linux-gnueabihf
The current snapcraft.yaml is using "on amd64 to armhf" to trigger
the install of the gcc-arm-linux-gnueabihf cross compiler. However
this does not work when building the snap in LP. This is reported
as LP:#1795430 and once this is fixed we can revert this change.

This PR works around this issue by simply using "on amd64" without
the "to armhf". This is fine because the snapcraft.yaml currently
only supports building on amd64 to armhf anyway, no other build
will work.