Remove the nobt/bt* split on the boot partition
This was one part of the boot configuration migration implemented in
Eoan that didn't quite work out as planned. The new pibootctl tool
doesn't require the extra complexity in all the configuration files -
it's enough to just have the config/syscfg/usercfg split without going
further into the bluetooth/no-bluetooth split.

Also use the serial0 alias in cmdline.txt; we should've been using this
all along. The firmware re-writes it to the correct UART.