snapcraft.yaml: support cross-compiling w/o --target-arch
Compiling with --target-arch is no longer supported by snapcraft, so instead
because this branch will only ever build for armhf, we can detect if we are
doing a native or cross-compilation build and just set the arch triplet
ourselves accordingly.

This also means we need to switch from using snapcraft's on: to: advanced
grammar for build-packages to just using try: which will work both natively and
when cross-compiling because failures trying to install these packages natively
will just get ignored.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>