.travis.yml: build the snap via lxd
snapcraft 4.0 enables us to build core20 images with lxd, meaning we can
properly test the snap out. Travis doesn't yet support focal, so building the
snap on destructively on bionic doesn't correctly test anything.

snapcraft 4.0 also doesn't support building as root, so instead we build as the
user, but ensure that the CI user is in the lxd group (because it is not by 
default it seems).

Finally, note that this results in a cross-compile build, which is also 
currently broken because it will not use the snappy-dev PPA, but it will at 
least use the correct arch packages. We will fix it to use the snappy-dev PPA w/
a future commit.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>