sources.list.cross: rename to cross-sources.list
This eliminates warning when cross-compiling such as

N: Ignoring file 'sources.list.cross' in directory '/home/user/git/pi-gadget/parts/gadget/src/stage/apt/multiverse.sources.list.d/' as it has an invalid filename extension

which happens because in the cross-compiling case we copy files from the helpers
and the filename isn't recognized by apt as a valid source file. The message is
harmless, but not ideal when looking through logs as it can be confusing.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>