Don't store ethaddr, usbethaddr, or serial# in env
The original uboot environment uses "saveenv" which dumps the *entire*
environment, including values like serial#, ethaddr, and usbethaddr (the
latter two are derived from the former) into the environment. If the SD
card is then switched to another Pi, U-Boot will read those values from
the environment and overwrite the board serial number, and ethernet MAC
address in the device-tree before passing it to the Linux kernel.

This modification introduces an "export_env" value which can be called
(with run) instead of saveenv to export the environment excluding those
values. The list of values exported is in the new "export_vars" value.
This is extensive because the environment is effectively the boot
*script* in addition to the boot *state*. The result has been tested
with armhf and arm64 builds on a 3B+, switching the card to a 3B and
confirming that the MAC address changes.

LP: #1900879