Enable then dwc2 USB host driver on arm64
BugLink: https://bugs.launchpad.net/bugs/1900665

The default non-upstream dwc_otg driver supports FIQ interrupt handling
(presumable for better performance) but lacks the maintenance and bug
fixing of the vanilla upstream dwc2 driver. Turns out that FIQ handling is
disabled in the dwc_otg driver on arm64 so it makes little sense to use
that driver on arm64.

In fact, upstream raspberrypi recommends to use dwc2 on arm64, so let's
do that.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>