overlay_map.dtb must live under overlays/
By default it's winding up in the root of the boot partition. This was
already patched in flash-kernel (LP: #1922779) and so it's largely gone
unnoticed since flash-kernel is implicitly run during the desktop setup,
however people installing desktops onto the server image may still be
affected (up to the first point where they run flash-kernel, or cause it
to be run by, e.g., a kernel upgrade).