Fix dtparam ordering
Urgh -- dtparam's apply to the last dtoverlay before them. However,
we've got a bunch of dtparams that apply to the *base* overlay.
Theoretically one can use "dtoverlay=" to reset to the base overlay, but
if it's the *first* "dtoverlay=" statement it also disables auto-loading
of HAT overlays (which we don't want).

So the only option is to ensure our base overlay dtparam items come
first (and add a warning in the file to avoid user confusion). But that
also means the [cm4] section potentially gets in the way of includes so
we need to reset to [all] afterwards.