Specify the datasource as "local"
Following discussion in LP: #1991261, it appears the "nocloud"
datasource doesn't implicitly specify itself as "local". This means that
aspects of "user-data" (like the set-hostname module) aren't applied
prior to the network being brought up, which in turn results in certain
effects (like DHCP not seeing the "final" client ID of the machine on
the first boot).

After testing to ensure this doesn't affect remote seeding (which is now
documented in README), this change is made to ensure user-data is fully
applied prior to the network being activated on first boot.