Used standard netplan config, and add regdom
It turns out that standard netplan configuration (including the
top-level "network" key) is permissible in network-config. The
cloud-init documentation is just rather mis-leading on the subject.

Furthermore, the netplan version now shipping in kinetic supports
specified on the wifi regulatory domain in the netplan configuration.
This is planned to be back-ported to jammy but this hasn't landed yet so
this commit limits the change to kinetic onwards.