Simplify the gadget logic on this branch dedicated to UC24/arm64
A lot of tests in the Makefile are useless as we maintain a dedicated
branch per UC version and architecture. The commit simplifies the gadget
logic, and get rid of the Makefile by migrating everything into the
snapcraft.yaml file.

The bootloader configuration file and the kernel command line have been
merged in a single file.

Helpers for cross-compilation are not needed, thus removed.