Fix handling of overlay_map, hat_map and README
The README file (which is used by the dtoverlay and dtparam utilities)
should appear under the overlays directory. This is not currently
happening with the gadget, although a flash-kernel run later fixes this.
At the same time, we need to ensure the change doesn't affect builds of
the gadget for legacy releases (particularly jammy and prior which still
have armhf variants)