Remove the deprecated z3fold pool allocator
This is a minor change in that z3fold no longer exists in the (proposed)
questing kernel, so the boot falls back to zsmalloc (the new default)
anyway. However, it does leave an error in the kernel log on every boot,
which we should avoid ideally.

The corresponding change for initramfs-tools' modules list has already
been published in ubuntu-raspi-settings.

LP: #2125496