Ensure corrupted initramfs also falls back in A/B boot
During further ISO tests with the proposed kernel in questing, we
discovered that the hardware watchdog is insufficient in cases where the
kernel is still running but has panicked because (for example) the
initramfs is corrupt.

Adds panic=10 to the default kernel configuration to ensure that in this
circumstance, the machine still resets providing an opportunity for the
bootloader to fallback to the "known good" configuration.